// frontend/src/features/node/editor/types/styles.ts
export enum TextAlignmentEnum {
	Start = 'start',
	End = 'end',
	Center = 'center',
	Justify = 'justify',
}

export enum JustificationEnum {
	Top = 'flex-start',
	Middle = 'center',
	Bottom = 'flex-end',
	Baseline = 'baseline',
}

export enum BorderRadiusEnum {
	None = 0,
	XS = 2,
	SM = 4,
	MD = 6,
	LG = 8,
	XL = 12,
}

export enum TextFontWeightEnum {
	Light = 300,
	Regular = 400,
	Medium = 500,
	SemiBold = 600,
	Bold = 700,
	ExtraBold = 800,
}

export enum TextFontSizeEnum {
	S = 12,
	M = 14,
	L = 16,
	XL = 18,
	XXL = 20,
}